fix(adapters/qwen-local): address Greptile review comments from PR #1490
- execute.ts: pass runtimeEnv (not env) to runChildProcess so subprocess
  receives the PATH-ensured env matching ensureCommandResolvable's
  pre-flight check. (runChildProcess does merge process.env internally,
  so the change is functionally a no-op, but it makes intent explicit.)
- test.ts: same change on the hello probe's runChildProcess call.
- shared/utils.ts: extract asRecord / asString / asNumber /
  stringifyUnknown / errorText helpers (previously duplicated across
  cli/format-event.ts and ui/parse-stdout.ts) into a shared module.

Addresses all three P2 inline comments on paperclipai/paperclip#1490.
